Bug 560815 - PRIVATE_TO_JSVAL(NULL) yields warning: NULL used in arithmetic with gcc 4.2.1. r=jorendorff.
UltraBlame original commit: 450522cbb138f0641b9c2c741bd959f442eb5156

+++ js/src/jsapi.h
@@ -270,7 +270,7 @@ BOOLEAN_TO_JSVAL(JSBool b)
 
 
 #define JSVAL_TO_PRIVATE(v)     ((void *)((v) & ~JSVAL_INT))
-#define PRIVATE_TO_JSVAL(p)     ((jsval)(p) | JSVAL_INT)
+#define PRIVATE_TO_JSVAL(p)     ((jsval)(ptrdiff_t)(p) | JSVAL_INT)
 
 
 #define JSPROP_ENUMERATE        0x01    /* property is visible to for/in loop */
